DB-10786 fixed remote_it

+++ hbase_sql/src/test/java/com/splicemachine/derby/utils/VacuumIT.java
@@ -194,7 +194,7 @@ public class VacuumIT extends SpliceUnitTest{
 
     @Test
     public void testVacuumDoesNotDeleteTablePossiblyInUse() throws Exception {
-        int oldest = SpliceUnitTest.getOldestActiveTransaction(methodWatcher);
+        long oldest = SpliceUnitTest.getOldestActiveTransaction(methodWatcher);
         LOG.info("VacuumIT: oldest: " + oldest + "\n");
         Connection connection = spliceClassWatcher.getOrCreateConnection();
